pyproject.toml
[project]
name = "bentovlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-index>=0.12.2",
    "pillow>=11.0.0",
    "bentoml>=1.3.14",
]

[tool.bentoml.build]
service = "service:VLMService"

[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev.cmd = "bentoml serve"
dev.env_file = ".env"
deploy = "bentoml deploy ."
